profile-manager/requirements.txt
fastapi==0.89.1
httpx==0.23.3
pydantic==1.10.4
pymongo==4.3.3
redis==4.4.2
requests-cache==0.9.8
requests==2.28.2
square-auth==0.0.14
square-elk-json-formatter==0.0.3
square-skill-api==0.0.47
uvicorn==0.20.0
square-model-client==0.0.10
pyyaml==6.0
